requirements.txt
dvc>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
great-expectations>=0.18.0
pytest>=7.0.0
yfinance>=0.2.0
sec-edgar-downloader>=5.0.0
transformers>=4.0.0
torch>=2.0.0
lightgbm>=4.0.0
mlflow>=2.0.0
pyarrow>=14.0.0
scikit-learn>=1.3.0
